feat + test: 'Added logic so non-admin users'
- 'Added logic so non-admin users only see their own data' página 34 (37)

- Es has-authority no has-role

- Mejoras en PointResource: métodos get, update y delete

- Mejoras: los usuarios no administradores no pueden ver a los otros usuarios

TEST:
- Ya funciona el test createPreference de PreferenceResourceIntTest
- 2 enfoques: PreferenceResourceIntTest y PointResourceIntTest
- REFERENCIAS:
	. SecurityUtilsUnitTest
	. https://github.com/spring-projects/spring-data-examples/blob/master/jpa/security/src/test/java/example/springdata/jpa/security/SecurityIntegrationTests.java
	. libro página 35 (38)

+++ src/main/java/com/desprogramar/jhipster/web/rest/PointResource.java
@@ -3,11 +3,15 @@ package com.desprogramar.jhipster.web.rest;
 import com.codahale.metrics.annotation.Timed;
 import com.desprogramar.jhipster.domain.Point;
 import com.desprogramar.jhipster.repository.PointRepository;
+import com.desprogramar.jhipster.repository.UserRepository;
 import com.desprogramar.jhipster.repository.search.PointSearchRepository;
+import com.desprogramar.jhipster.security.AuthoritiesConstants;
+import com.desprogramar.jhipster.security.SecurityUtils;
 import com.desprogramar.jhipster.web.rest.util.HeaderUtil;
 import com.desprogramar.jhipster.web.rest.util.PaginationUtil;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpHeaders;
@@ -35,13 +39,16 @@ import static org.elasticsearch.index.query.QueryBuilders.*;
 public class PointResource {
 
     private final Logger log = LoggerFactory.getLogger(PointResource.class);
-        
+
     @Inject
     private PointRepository pointRepository;
-    
+
     @Inject
     private PointSearchRepository pointSearchRepository;
-    
+
+    @Autowired
+    private UserRepository userRepository;
+
     /**
      * POST  /points -> Create a new point.
      */
@@ -53,6 +60,10 @@ public class PointResource {
         log.debug("REST request to save Point : {}", point);
         if (point.getId() != null) {
             return ResponseEntity.badRequest().headers(HeaderUtil.createFailureAlert("point", "idexists", "A new point cannot already have an ID")).body(null);
+        }
+        if(!SecurityUtils.isCurrentUserInRole(AuthoritiesConstants.ADMIN) || point.getUser() == null) {
+            log.debug("No user passed in, using current user: {}", SecurityUtils.getCurrentUserLogin());
+            point.setUser(userRepository.findOneByLogin(SecurityUtils.getCurrentUserLogin()).get());
         }
         Point result = pointRepository.save(point);
         pointSearchRepository.save(result);
@@ -73,6 +84,10 @@ public class PointResource {
         if (point.getId() == null) {
             return createPoint(point);
         }
+        if(!SecurityUtils.isCurrentUserInRole(AuthoritiesConstants.ADMIN) || point.getUser() == null) {
+            log.debug("No user passed in, using current user: {}", SecurityUtils.getCurrentUserLogin());
+            point.setUser(userRepository.findOneByLogin(SecurityUtils.getCurrentUserLogin()).get());
+        }
         Point result = pointRepository.save(point);
         pointSearchRepository.save(result);
         return ResponseEntity.ok()
@@ -90,7 +105,12 @@ public class PointResource {
     public ResponseEntity<List<Point>> getAllPoints(Pageable pageable)
         throws URISyntaxException {
         log.debug("REST request to get a page of Points");
-        Page<Point> page = pointRepository.findAll(pageable); 
+        Page<Point> page;
+        if (SecurityUtils.isCurrentUserInRole(AuthoritiesConstants.ADMIN)) {
+            page = pointRepository.findAllByOrderByDateDesc(pageable);
+        } else {
+            page = pointRepository.findAllByUserIsCurrentUser(pageable);
+        }
         HttpHeaders headers = PaginationUtil.generatePaginationHttpHeaders(page, "/api/points");
         return new ResponseEntity<>(page.getContent(), headers, HttpStatus.OK);
     }
@@ -105,6 +125,12 @@ public class PointResource {
     public ResponseEntity<Point> getPoint(@PathVariable Long id) {
         log.debug("REST request to get Point : {}", id);
         Point point = pointRepository.findOne(id);
+        if (point != null &&
+            !SecurityUtils.isCurrentUserInRole(AuthoritiesConstants.ADMIN) &&
+            !point.getUser().getLogin().equals(SecurityUtils.getCurrentUserLogin())) {
+                return ResponseEntity.badRequest().headers(HeaderUtil.createFailureAlert("point", "idNoExists", "El ID '"+id+"' no existe!")).body(null);
+                // return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
         return Optional.ofNullable(point)
             .map(result -> new ResponseEntity<>(
                 result,
@@ -121,6 +147,13 @@ public class PointResource {
     @Timed
     public ResponseEntity<Void> deletePoint(@PathVariable Long id) {
         log.debug("REST request to delete Point : {}", id);
+        Point point = pointRepository.findOne(id);
+        if (point != null &&
+            !SecurityUtils.isCurrentUserInRole(AuthoritiesConstants.ADMIN) &&
+            !point.getUser().getLogin().equals(SecurityUtils.getCurrentUserLogin())) {
+            return ResponseEntity.badRequest().headers(HeaderUtil.createFailureAlert("point", "idNoExists", "El ID '"+id+"' no existe!")).build();
+            // return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
         pointRepository.delete(id);
         pointSearchRepository.delete(id);
         return ResponseEntity.ok().headers(HeaderUtil.createEntityDeletionAlert("point", id.toString())).build();
